src/requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pefile==2019.4.18
Pillow==7.2.0
pycryptodome==3.9.8
PyInstaller==3.6
pywin32-ctypes==0.2.0
requests==2.24.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
